refactor: solve lint problems
