+++ apps/auth-service/src/users-informations/users-informations.service.ts
@@ -1,4 +1,9 @@
-import { forwardRef, Inject, Injectable, NotFoundException } from '@nestjs/common';
+import {
+  forwardRef,
+  Inject,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import {
   UserInformation,
   UserInformationDocument,
