+++ apps/auth-service/src/users/dto/update-user.dto.ts
@@ -1,7 +1,6 @@
-import { IsEmail, IsOptional, IsString, MinLength } from "class-validator";
+import { IsEmail, IsOptional, IsString, MinLength } from 'class-validator';
 
 export class UpdateUserDto {
-
   @IsString()
   @MinLength(1)
   @IsOptional()
