+++ apps/engage-hub-service/src/routines_exercises/entities/routine-exercise.entity.ts
@@ -23,10 +23,9 @@ export class RoutineExercise {
     name: 'status',
     type: 'enum',
     enum: ['not started', 'in progress', 'completed'],
-    default: 'not started', 
+    default: 'not started',
   })
   status: 'not started' | 'in progress' | 'completed';
-
 
   @ManyToOne(() => Routine, (routine) => routine.exercises, { cascade: true })
   routine: Routine;
