+++ apps/frontend/app/utils/sessionTokenAccessor.ts
@@ -1,5 +1,5 @@
 import getServerSession from "next-auth";
-import { authOptions } from '../api/auth/[...nextauth]/route'
+import { authOptions } from "@/app/utils/authOptions";
 import { decrypt } from "./encryption";
 
 export async function getAccessToken() {
